rename directory and modify paths
